// src/app/models/DichVu.ts
export interface DichVu{
  idDoc?: string;
  id?: number;
  imgDichVu?: string;
  maNhomDV?: string;
  tenDV?: string;
  camKet?: string[];
  luuY?: string[];
  moTa?: string[];
  danhSachSuaChua?: string[];
  danhSachDV?: string[];
  lyDoCanSua?: string[];
  ghiChuSuDung?: string[];
  tieuChiDV?: string[];
  quyTrinhTiepNhan?: string[];
  quyTrinhBD?: string[];
  luuYSuDung?: string[];
}